refactor: move starfield generation to planetaryrenderer
move sprite generation to planetary renderer as well
delete sprite generation file
