+++ src/render/PlanetaryRenderer.ts
@@ -1,4 +1,5 @@
-import * as THREE from "three";
+import * as THREE from 'three';
+import SimplexNoise from 'simplex-noise';
 
 let earthNormalTexture: THREE.Texture;
 
@@ -26,5 +27,43 @@ export function createSunMesh(mass: number): THREE.Group{
     sunGroup.add(new THREE.Mesh(geometry, material));
     sunGroup.add(light);
     return sunGroup;
+}
+
+export function createStarField(seed: number, totalNebulas: number): THREE.Group{
+    const spriteGroup = new THREE.Group();
+    const textureLoader = new THREE.TextureLoader();
+
+    const starMaterial = new THREE.SpriteMaterial( { map: textureLoader.load( 'assets/star.png' ) } );
+    const sprites = generateSprites( [starMaterial], seed);
+    spriteGroup.add( ... sprites );
+
+    const nebulaMaterials: THREE.SpriteMaterial[] = [];
+    for (let i = 1; i <= totalNebulas; i++)
+    nebulaMaterials.push(new THREE.SpriteMaterial( { map: textureLoader.load( 'assets/nebula' + i + '.png' ) } ));
+    const nebulas = generateSprites(nebulaMaterials, seed, 20, 5);
+    spriteGroup.add( ... nebulas );
+    return spriteGroup;
+}
+
+function generateSprites(spriteMaterials: THREE.SpriteMaterial[], seed?: number, totalSprites = 4000, scale = 1, distance = 3000, noiseScale = .005): THREE.Sprite[] {
+    const perlin = new SimplexNoise(seed);
+
+    const sprites: THREE.Sprite[] = [];
+    
+    while (sprites.length < totalSprites) {
+        const starPos = (new THREE.Vector3()).randomDirection().multiplyScalar(distance);
+
+        const starProb = (perlin.noise3D(starPos.x * noiseScale, starPos.y * noiseScale, starPos.z * noiseScale) + 1) / 2;
+
+        if (starProb > Math.random() * .7 + .3) {
+            const sprite = new THREE.Sprite(spriteMaterials[Math.floor(Math.random() * spriteMaterials.length)]);
+            sprite.position.set(starPos.x, starPos.y, starPos.z);
+            const size = (Math.pow(2, Math.random() * 7 + 3) / 20 + 15) * scale;
+            sprite.scale.set(size, size, size);
+            sprites.push(sprite);  
+        }
+    }
+
+    return sprites;
 }
